requirements.txt
openpyxl>=3.1.2
pandas>=2.0.0
line-bot-sdk>=3.5.0
flask>=3.0.0
python-dotenv>=1.0.0
